changing function location to atoms.affine.upper_tri
